requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
audioread==2.1.8
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
comtypes==1.1.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
distlib==0.3.0
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.3.3
graphviz==0.13.2
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.3.1
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
jieba3k==0.35.1
Jinja2==2.11.1
joblib==0.14.1
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mock==4.0.1
newspaper3k==0.2.8
nltk==3.4.5
numba==0.48.0
numpy==1.18.1
pandas==1.0.2
parso==0.6.2
pdf2image==1.12.1
PDPbox==0.2.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
plyer==1.4.2
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.6.1
pyler==0.2.0
pyparsing==2.4.6
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
requests-file==1.4.3
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
shap==0.35.0
six==1.14.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.0
tabulate==0.8.7
termcolor==1.1.0
tesseract==0.1.3
tinysegmenter==0.3
tldextract==2.2.2
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
wcwidth==0.1.8
Werkzeug==1.0.0
wincertstore==0.2
word2number==1.1
WTForms==2.2.1
zipp==3.1.0